@@ -2,18 +2,16 @@ package dev.alexmaycon.bucketservice.hook;
 
 import dev.alexmaycon.bucketservice.config.ServiceConfiguration;
 import dev.alexmaycon.bucketservice.hook.model.Hook;
-import org.glassfish.jersey.client.ClientConfig;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.core.JobExecution;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.ws.rs.client.Client;
-import javax.ws.rs.client.ClientBuilder;
-import javax.ws.rs.client.Entity;
-import javax.ws.rs.client.WebTarget;
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.client.Client;
+import jakarta.ws.rs.client.ClientBuilder;
+import jakarta.ws.rs.client.Entity;
+import jakarta.ws.rs.client.WebTarget;
+import jakarta.ws.rs.core.Response;
 import java.util.HashMap;
 
 @Component
@@ -22,7 +20,6 @@ public class HookClientRequest {
     private static final Logger logger = LoggerFactory.getLogger(HookClientRequest.class);
     private final ServiceConfiguration serviceConfiguration;
 
-    @Autowired
     public HookClientRequest(ServiceConfiguration serviceConfiguration) {
         this.serviceConfiguration = serviceConfiguration;
     }
@@ -36,9 +33,7 @@ public class HookClientRequest {
 
             final String contentType = serviceConfiguration.getService().getHookContentType();
 
-            ClientConfig config = new ClientConfig();
-
-            Client client = ClientBuilder.newClient(config);
+            Client client = ClientBuilder.newClient();
 
             WebTarget target = client.target(hook);
 
